26/12/2021

+++ Api.Api/Controllers/SpeakersController.cs
@@ -2,6 +2,7 @@
 using System.Net;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
+using AutoMapper;
 using Api.Application.Dtos.Speaker;
 using Api.Application.Interfaces;
 
@@ -13,9 +14,13 @@ namespace Api.Api.Controllers
     public class SpeakersController : ControllerBase
     {
         private ISpeakerService _service { get; set; }
-        public SpeakersController(ISpeakerService service)
+        private IWebHostEnvironment _hostEnvironment;
+        private readonly IMapper _mapper;
+        public SpeakersController(ISpeakerService service, IWebHostEnvironment hostEnvironment, IMapper mapper)
         {
             _service = service;
+            _hostEnvironment = hostEnvironment;
+            _mapper = mapper;
         }
 
         // [Authorize("Bearer")]
@@ -32,20 +37,6 @@ namespace Api.Api.Controllers
             }
         }
 
-        [HttpGet]
-        [Route("{skip}/{take}")]
-        public async Task<ActionResult> GetAllPage([FromRoute] int skip, [FromRoute] int take)
-        {
-            try
-            {
-                return Ok(await _service.GetAllPage(skip, take));
-            }
-            catch (Exception e)
-            {
-                return StatusCode((int)HttpStatusCode.InternalServerError, e.Message);
-            }
-        }
-
         // [Authorize("Bearer")]
         [HttpGet]
         [Route("{id}", Name = "GetSpeakerWithId")]
@@ -53,12 +44,12 @@ namespace Api.Api.Controllers
         {
             try
             {
-                var result = await _service.Get(id);
+                var result = await _service.GetById(id);
                 if (result == null)
                 {
                     return NotFound($"Pesquisa não obteve êxito com Id: {id}");
                 }
-                return Ok(await _service.Get(id));
+                return Ok(await _service.GetById(id));
 
             }
             catch (ArgumentException e)
@@ -116,17 +107,79 @@ namespace Api.Api.Controllers
         {
             try
             {
-                var result = await _service.Get(id);
+                var result = await _service.GetById(id);
                 if (result == null)
                 {
                     return NotFound($"Deleção não obteve êxito com Id: {id}");
                 }
-                return Ok(await _service.Delete(id));
+                if (await _service.Delete(id))
+                {
+                    DeleteImage(result.SpeakerImage);
+                    return Ok(await _service.Delete(id));
+                }
+                else
+                {
+                    throw new Exception($"Erro ao deletar evento {result.SpeakerName}");
+                }
             }
             catch (ArgumentException ex)
             {
                 return StatusCode((int)System.Net.HttpStatusCode.InternalServerError, ex.Message);
             }
+        }
+
+
+        //////////////////////  Updaload de imagem
+        [HttpPost("updalod-image/{speakerId}")]
+        public async Task<ActionResult> UploadImage(Guid speakerId)
+        {
+
+            try
+            {
+                var result = await _service.GetById(speakerId);
+                if (result is null) return NoContent();
+
+                var file = Request.Form.Files[0];
+                Console.WriteLine(file);
+                if (file.Length > 0)
+                {
+                    DeleteImage(result.SpeakerImage);
+                    result.SpeakerImage = await SaveImage(file);
+                }
+                var resultMap = _mapper.Map<SpeakerUpdateDto>(result);
+                var speakerReturn = await _service.Put(resultMap);
+
+                return Ok(speakerReturn);
+            }
+            catch (ArgumentException ex)
+            {
+                throw new ArgumentException("Erro ao salvar imagem", ex.Message);
+            }
+        }
+
+        [NonAction]
+        public async Task<string> SaveImage(IFormFile imageFile) //Save image on folder Resourses
+        {
+            string imageName = new String(Path.GetFileNameWithoutExtension(imageFile.FileName).Take(10).ToArray()).Replace(' ', '-');
+
+            imageName = $"{imageName}{DateTime.UtcNow.ToString("yymmssfff")}{Path.GetExtension(imageFile.FileName)}";
+
+            var imagePath = Path.Combine(_hostEnvironment.ContentRootPath, @"resources/images", imageName);
+
+            using (var fileStream = new FileStream(imagePath, FileMode.Create))
+            {
+                await imageFile.CopyToAsync(fileStream);
+            }
+
+            return imageName;
+        }
+
+        [NonAction]  //delete the image from folder Resourses
+        public void DeleteImage(string imageName)
+        {
+            var imagePath = Path.Combine(_hostEnvironment.ContentRootPath, @"resources/images", imageName);
+            if (System.IO.File.Exists(imagePath))
+                System.IO.File.Delete(imagePath);
         }
     }
 }
